+++ game/game.js
@@ -1,23 +1,30 @@
 'use strict';
 
+import * as PIXI from 'pixi.js';
+import { keyboard } from './keyboard';
+
 const renderer =
   PIXI.autoDetectRenderer(window.innerWidth,
     window.innerHeight,
-    { backgroundColor: 0x1099bb });
+    { backgroundColor: 0x000000 });
 document.body.appendChild(renderer.view);
 
 // create the root of the scene graph
 const stage = new PIXI.Container();
 
+const bunnyTexture = PIXI.Texture.fromImage('assets/bunny.png');
+const bulletTexture = PIXI.Texture.fromImage('assets/bullet.png');
+const mapTexture = PIXI.Texture.fromImage('assets/levels/baseMap.png');
+
 // create a background
 const dungeon = new PIXI.Sprite(mapTexture);
 dungeon.scale.x = 2.7;
 dungeon.scale.y = 2.7;
+
 stage.addChild(dungeon);
 
-
 // create a new Sprite using the texture
-const player = new PIXI.Sprite(bunnyTexture);
+export const player = new PIXI.Sprite(bunnyTexture);
 
 // move the sprite to the center of the screen
 player.position.x = renderer.width / 2;
@@ -111,6 +118,26 @@ function shoot(mX, mY) {
 
   //ScreenShake(10);
 }
+
+// character control
+const linearSpeed = 7;
+
+const left = keyboard('a'),
+  up = keyboard('w'),
+  right = keyboard('d'),
+  down = keyboard('s');
+
+left.press = () => player.vx -= linearSpeed;
+left.release = () => player.vx += linearSpeed;
+
+up.press = () => player.vy += linearSpeed;
+up.release = () => player.vy -= linearSpeed;
+
+right.press = () => player.vx += linearSpeed;
+right.release = () => player.vx -= linearSpeed;
+
+down.press = () => player.vy -= linearSpeed;
+down.release = () => player.vy += linearSpeed;
 
 function ScreenShake(shake) {
   dungeon.position.x += shake;
