--- game/textures.js
@@ -1,5 +1,0 @@
-'use strict';
-
-const bunnyTexture = PIXI.Texture.fromImage('assets/bunny.png');
-const bulletTexture = PIXI.Texture.fromImage('assets/bullet.png');
-const mapTexture = PIXI.Texture.fromImage('assets/levels/baseMap.png');
