+++ index.html
@@ -5,10 +5,8 @@
     <title>Rage Party</title>
 </head>
 <body>
-    <script src="scripts/pixi.min.js"></script>
+<!--    <script src="scripts/pixi.min.js"></script>-->
     <script src="game/keyboard.js"></script>
-    <script src="game/characterControll.js"></script>
-    <script src="game/textures.js"></script>
     <script src="game/wallCollisions.js"></script>
     <script src="game/game.js"></script>
 </body>
