+++ package.json
@@ -20,10 +20,18 @@
   },
   "homepage": "https://github.com/vermi4elli/Rage-Party#readme",
   "devDependencies": {
-    "eslint": "^7.10.0"
+    "copy-webpack-plugin": "^6.2.1",
+    "eslint": "^7.10.0",
+    "html-webpack-plugin": "^4.5.0",
+    "imagemin-webpack-plugin": "^2.4.2",
+    "webpack": "^5.2.0",
+    "webpack-cli": "^4.1.0",
+    "webpack-dev-server": "^3.11.0",
+    "webpack-merge": "^5.2.0"
   },
   "dependencies": {
     "fastify": "^3.5.1",
-    "fastify-static": "^3.2.1"
+    "fastify-static": "^3.2.1",
+    "pixi.js": "^5.3.3"
   }
 }
